+++ vector/src/main/java/im/vector/app/features/themes/ActivityOtherThemes.kt
@@ -31,6 +31,11 @@ sealed class ActivityOtherThemes(@StyleRes val dark: Int,
             R.style.AppTheme_Black
     )
 
+    object Launcher : ActivityOtherThemes(
+            R.style.AppTheme_Launcher,
+            R.style.AppTheme_Launcher
+    )
+
     object AttachmentsPreview : ActivityOtherThemes(
             R.style.AppTheme_AttachmentsPreview,
             R.style.AppTheme_AttachmentsPreview